Gerar UUID 89

+++ Codigo/ModuloDeSoftware/WebAPI/SalasUfsWeb/Controllers/EquipamentoController.cs
@@ -50,7 +50,7 @@ namespace SalasUfsWeb.Controllers
         {
             var equipamentosModel = _equipamentoService.GetAll();
             List<EquipamentoViewModel> equipamentos = new List<EquipamentoViewModel>();
-            equipamentosModel.ForEach(e => equipamentos.Add(new EquipamentoViewModel { EquipamentoModel = e, SalaModel = _salaService.GetById(e.Sala), HardwareDeSalaModel = _hardwareDeSalaService.GetById(e.HardwareDeSala) }));
+            equipamentosModel.ForEach(e => equipamentos.Add(new EquipamentoViewModel { EquipamentoModel = e, SalaModel = _salaService.GetById(e.Sala), HardwareDeSalaModel = _hardwareDeSalaService.GetById((int) e.HardwareDeSala) }));
 
             return View(equipamentos);
         }
@@ -155,7 +155,7 @@ namespace SalasUfsWeb.Controllers
                 BlocoModel = bloco,
                 OrganizacaoModel = _organizacaoService.GetById(bloco.OrganizacaoId),
                 Codigos = codigosView,
-                HardwareDeSalaModel = _hardwareDeSalaService.GetById(equipamento.HardwareDeSala)
+                HardwareDeSalaModel = _hardwareDeSalaService.GetById((int) equipamento.HardwareDeSala)
             };
 
             string[] tiposEquipamento = { EquipamentoModel.TIPO_CONDICIONADOR, EquipamentoModel.TIPO_LUZES };
